+++ elmyra.ip.access.epo/CHANGES.rst
@@ -6,6 +6,7 @@ development
 ===========
 - ui: refactor query builder to separate component
 - ui: introduce field-based query builder (comfort flavor)
+- ui: update cql field chooser for DEPATISnet
 
 0.33.0
 ======